requirements.txt
django==5.0
python-telegram-bot==13.15
environs==11.0.0
django-phonenumber-field[phonenumberslite]==8.0.0
pillow==10.4.0
